solid-srp/src/main/java/ca/ulaval/glo4002/solid_srp/MovieTicketService.java
package ca.ulaval.glo4002.solid_srp;

public class MovieTicketService {

    private TicketPrinter printer;
    private TicketFactory factory;

    public MovieTicketService(TicketFactory factory, TicketPrinter printer) {
        this.factory = factory;
        this.printer = printer;
    }

    // TODO add : a flag to know if this is a student. Apply 10%.
    // TODO add : a flag to know whether the ticket should be printed or not.
    public void generateMovieTicket(int personAge, boolean isStudent, boolean printTicket) {
        MovieTicket ticket = factory.createTicket(personAge, isStudent);
        if (printTicket) {
            print(ticket);
        }
    }

    private void print(MovieTicket ticket) {
        printer.print(ticket);
    }
}
